Adaptive Best-Subset Selection via Splicing

Adaptive best-subset selection for regression, binary classification and censored-response modeling in polynomial times.

# S3 method for default
abess(
  x,
  y,
  family = c("gaussian", "binomial", "poisson", "cox", "mgaussian", "multinomial"),
  tune.path = c("sequence", "gsection"),
  tune.type = c("gic", "ebic", "bic", "aic", "cv"),
  weight = rep(1, nrow(x)),
  normalize = NULL,
  c.max = 2,
  support.size = NULL,
  gs.range = NULL,
  always.include = NULL,
  max.splicing.iter = 20,
  screening.num = NULL,
  warm.start = TRUE,
  nfolds = 5,
  cov.update = TRUE,
  newton = c("exact", "approx"),
  newton.thresh = 1e-06,
  max.newton.iter = NULL,
  early.stop = FALSE,
  num.threads = 0,
  seed = 1,
  ...
)

# S3 method for formula
abess(formula, data, subset, na.action, ...)

Arguments

x
Input matrix, of dimension \(n \times p\); each row is an observation vector and each column is a predictor/feature/variable.

y
The response variable, of n observations. For family = "binomial" should have two levels. For family="poisson", y should be a vector with positive integer. For family = "cox", y should be a two-column matrix with columns named time and status. For family = "mgaussian", y should be a matrix of quantitative responses. For family = "multinomial", y should be a factor of at least three levels. Note that, for either "binomial" or "multinomial", if y is presented as a numerical vector, it will be coerced into a factor.

family
One of the following models: "gaussian" (continuous response), "binomial" (binary response), "poisson" (non-negative count), "cox" (left-censored response), "mgaussian" (multivariate continuous response). Depending on the response. Any unambiguous substring can be given.

tune.path
The method to be used to select the optimal support size. For method = "sequence", we solve the best subset selection problem for each size in support.size. For method = "gsection", we solve the best subset selection problem with support size ranged in gs.range, where the specific support size to be considered is determined by golden section.

tune.type
The type of criterion for choosing the support size. Available options are "gic", "ebic", "bic", "aic" and "cv". Default is "gic".

weight
Observation weights. Default is 1 for each observation.

normalize
Options for normalization. normalize = 0 for no normalization. normalize = 1 for subtracting the mean of columns of x. normalize = 2 for scaling the columns of x to have \(\sqrt n\) norm. normalize = 3 for subtracting the means of the columns of x and y, and also normalizing the columns of x to have \(\sqrt n\) norm. If normalize = NULL, normalize will be set 1 for "gaussian", 2 for "binomial". Default is normalize = NULL.

c.max
an integer splicing size. Default is: c.max = 2.

support.size
An integer vector representing the alternative support sizes. Only used for method = "sequence". Default is 0:min(n, round(n/(log(log(n))log(p)))).

gs.range
A integer vector with two elements. The first element is the minimum model size considered by golden-section, the later one is the maximum one. Default is gs.range = c(1, min(n, round(n/(log(log(n))log(p))))).

always.include
An integer vector containing the indexes of variables that should always be included in the model.

max.splicing.iter
The maximum number of performing splicing algorithm. In most of the case, only a few times of splicing iteration can guarantee the convergence. Default is max.splicing.iter = 20.

screening.num
An integer number. Preserve screening.num number of predictors with the largest marginal maximum likelihood estimator before running algorithm.

warm.start
Whether to use the last solution as a warm start. Default is warm.start = TRUE.

nfolds
The number of folds in cross-validation. Default is nfolds = 5. For variables in the same group, they should be located in adjacent columns of x and their corresponding index in group.index should be the same. Denote the first group as 1, the second 2, etc. If you do not fit a model with a group structure, please set group.index = NULL. Default is NULL.

cov.update
A logical value only used for family = "gaussian". If cov.update = TRUE, use a covariance-based implementation; otherwise, a naive implementation. The naive method is more efficient than covariance-based method only when \(p >> n\). Default: cov.update = TRUE.

newton
A character specify the Newton's method for fitting generalized linear models, it should be either newton = "exact" or newton = "approx". If newton = "exact", then the exact hessian is used, while newton = "approx" uses diagonal entry of the hessian, and can be faster (especially when family = "cox").

newton.thresh
a numerica value for controlling positive convergence tolerance. The Newton's iterations converge when \(|dev - dev_{old}|/(|dev| + 0.1)<\) newton.thresh.

max.newton.iter
a integer giving the maximal number of Newton's iteration iterations. Default is max.newton.iter = 10 if newton = "exact", and max.newton.iter = 60 if newton = "approx".

early.stop
A boolean value decide whether early stoping. If early.stop = TRUE, algorithm will stop if the last tuning value less than the existing one. Default: early.stop = FALSE.

num.threads
An integer decide the number of threads to be concurrently used for cross-validation (i.e., tune.type = "cv"). If num.threads = 0, then all of available cores will be used. Default: num.threads = 0.

seed
Seed to be used to divide the sample into cross-validation folds. Default is seed = 1.

...
further arguments to be passed to or from methods.

formula
an object of class "formula": a symbolic description of the model to be fitted. The details of model specification are given in the "Details" section of "formula".

data
a data frame containing the variables in the formula.

subset
an optional vector specifying a subset of observations to be used.

na.action
a function which indicates what should happen when the data contain NAs. Defaults to getOption("na.action").

Value

A abess class object, which is a list with the following components:

beta

A \(p\)-by-length(support.size) matrix of coefficients for univariate family, stored in column format; while a list of length(support.size) coefficients matrix (with size \(p\)-by-ncol(y)) for multivariate family.

intercept

An intercept vector of length length(support.size) for univariate family; while a list of length(support.size) intercept vector (with size ncol(y)) for multivariate family.

dev

the deviance of length length(support.size).

tune.value

A value of tuning criterion of length length(support.size).

nobs

The number of sample used for training.

nvars

The number of variables used for training.

family

Type of the model.

tune.path

The path type for tuning parameters.

support.size

The actual support.size values used. Note that it is not necessary the same as the input if the later have non-integer values or duplicated values.

best.size

The best support size selected by the tuning value.

tune.type

The criterion type for tuning parameters.

tune.path

The strategy for tuning parameters.

screening.vars

The character vector specify the feature selected by feature screening. It would a empty character vector if screening.num = 0.

call

The original call to abess.

References

A polynomial algorithm for best-subset selection problem. Junxian Zhu, Canhong Wen, Jin Zhu, Heping Zhang, Xueqin Wang. Proceedings of the National Academy of Sciences Dec 2020, 117 (52) 33117-33123; DOI: 10.1073/pnas.2014241117

Sure independence screening for ultrahigh dimensional feature space. Fan, J. and Lv, J. (2008), Journal of the Royal Statistical Society: Series B (Statistical Methodology), 70: 849-911. https://doi.org/10.1111/j.1467-9868.2008.00674.x

See also

Author

Jin Zhu, Junxian Zhu, Canhong Wen, Heping Zhang, Xueqin Wang

Examples

n <- 100
p <- 20
support.size <- 3

################ linear model ################
dataset <- generate.data(n, p, support.size)
abess_fit <- abess(dataset[["x"]], dataset[["y"]])
## helpful generic functions:
print(abess_fit)

#> Call:
#> abess.default(x = dataset[["x"]], y = dataset[["y"]])
#> 
#>    support.size       dev      GIC
#> 1             0 17805.684 978.7273
#> 2             1 10163.270 927.2286
#> 3             2  4407.395 848.2540
#> 4             3  1842.745 765.6262
#> 5             4  1667.015 760.1791
#> 6             5  1594.016 760.2763
#> 7             6  1543.874 761.6551
#> 8             7  1518.615 764.5805
#> 9             8  1500.112 767.9297
#> 10            9  1486.034 771.5618
#> 11           10  1473.749 775.3067
#> 12           11  1460.411 778.9725
#> 13           12  1452.290 782.9899
#> 14           13  1444.224 787.0080
#> 15           14  1439.776 791.2745
#> 16           15  1437.522 795.6929
#> 17           16  1435.833 800.1504
#> 18           17  1434.539 804.6352
#> 19           18  1433.386 809.1298
#> 20           19  1432.351 813.6327
#> 21           20  1432.225 813.6239

coef(abess_fit, support.size = 3)

#> 21 x 1 sparse Matrix of class "dgCMatrix"
#>                     3
#> (intercept) -4.369861
#> x1           .       
#> x2           .       
#> x3           .       
#> x4           .       
#> x5           .       
#> x6          83.270800
#> x7           .       
#> x8           .       
#> x9           .       
#> x10          .       
#> x11          .       
#> x12          .       
#> x13          .       
#> x14         89.689944
#> x15          .       
#> x16          .       
#> x17          .       
#> x18         43.410846
#> x19          .       
#> x20          .       

predict(abess_fit, newx = dataset[["x"]][1:10, ], 
        support.size = c(3, 4))

#>                3           4
#>  [1,]  103.05157   91.524774
#>  [2,]   74.66998   85.359664
#>  [3,] -289.97309 -299.858907
#>  [4,]  -16.35758   -4.241221
#>  [5,]  171.80572  162.807112
#>  [6,]  126.58540  127.021354
#>  [7,] -197.24366 -207.134508
#>  [8,] -126.67823 -142.927367
#>  [9,]  -23.29128  -22.898065
#> [10,] -109.76937 -117.273626

str(extract(abess_fit))

#> List of 7
#>  $ beta        :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:4] 5 13 15 17
#>   .. ..@ p       : int [1:2] 0 4
#>   .. ..@ Dim     : int [1:2] 20 1
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : chr [1:20] "x1" "x2" "x3" "x4" ...
#>   .. .. ..$ : chr "4"
#>   .. ..@ x       : num [1:4] 82.4 89 -12.6 43.3
#>   .. ..@ factors : list()
#>  $ intercept   : num -4.77
#>  $ support.size: int 4
#>  $ support.vars: chr [1:4] "x6" "x14" "x16" "x18"
#>  $ support.beta: num [1:4] 82.4 89 -12.6 43.3
#>  $ dev         : num 1667
#>  $ tune.value  : num 760

deviance(abess_fit)

#>  [1] 17805.684 10163.270  4407.395  1842.745  1667.015  1594.016  1543.874
#>  [8]  1518.615  1500.112  1486.034  1473.749  1460.411  1452.290  1444.224
#> [15]  1439.776  1437.522  1435.833  1434.539  1433.386  1432.351  1432.225

################ logistic model ################
dataset <- generate.data(n, p, support.size, family = "binomial")
## allow cross-validation to tuning
abess_fit <- abess(dataset[["x"]], dataset[["y"]], 
                   family = "binomial", tune.type = "cv")
abess_fit

#> Call:
#> abess.default(x = dataset[["x"]], y = dataset[["y"]], family = "binomial", 
#>     tune.type = "cv")
#> 
#>    support.size       dev        cv
#> 1             0  68.59298 13.888785
#> 2             1  46.69364  9.467560
#> 3             2  22.31417  4.795569
#> 4             3  17.60334  5.453128
#> 5             4  16.86618  9.348024
#> 6             5  16.06971 14.488638
#> 7             6  15.26069 22.724323
#> 8             7  14.18543 33.229924
#> 9             8  12.91200 42.426394
#> 10            9  11.00353 56.632611
#> 11           10  10.04500 61.999518
#> 12           11  10.18355 66.914255
#> 13           12 173.00172 71.793735
#> 14           13 142.19030 80.653895
#> 15           14  88.00513 80.122038
#> 16           15 122.44452 82.172393
#> 17           16 141.91677 83.516764
#> 18           17 105.22047 83.446261
#> 19           18 109.50414 83.731671
#> 20           19 113.17618 84.137978
#> 21           20 154.73864 84.729315

################ Cox model ################
dataset <- generate.data(n, p, support.size, family = "cox")

#> censoring rate: 0.57 

abess_fit <- abess(dataset[["x"]], dataset[["y"]], 
                   family = "cox", tune.type = "cv", 
                   newton = "approx")
abess_fit

#> Call:
#> abess.default(x = dataset[["x"]], y = dataset[["y"]], family = "cox", 
#>     tune.type = "cv", newton = "approx")
#> 
#>    support.size      dev       cv
#> 1             0 186.9720 23.45101
#> 2             1 167.3854 20.53945
#> 3             2 142.9906 15.15529
#> 4             3 122.6477 12.43033
#> 5             4 119.1235 13.22219
#> 6             5 117.8076 13.72121
#> 7             6 116.9300 14.42269
#> 8             7 116.4218 14.29653
#> 9             8 115.2787 15.01824
#> 10            9 114.9203 15.39002
#> 11           10 114.5888 17.09378
#> 12           11 114.3221 18.53581
#> 13           12 114.2761 20.64661
#> 14           13 114.0693 21.82383
#> 15           14 113.9716 22.32343
#> 16           15 113.8485 21.99663
#> 17           16 113.7331 22.42952
#> 18           17 113.6946 22.48710
#> 19           18 113.6652 22.72920
#> 20           19 113.6552 22.69666
#> 21           20 113.6537 22.65670

################ Feature screening ################
p <- 1000
dataset <- generate.data(n, p, support.size)
abess_fit <- abess(dataset[["x"]], dataset[["y"]], 
                   screening.num = 100)
extract(abess_fit)

#> $beta
#> 1000 x 1 sparse Matrix of class "dgCMatrix"
#>       0
#> x1    .
#> x2    .
#> x3    .
#> x4    .
#> x5    .
#> x6    .
#> x7    .
#> x8    .
#> x9    .
#> x10   .
#> x11   .
#> x12   .
#> x13   .
#> x14   .
#> x15   .
#> x16   .
#> x17   .
#> x18   .
#> x19   .
#> x20   .
#> x21   .
#> x22   .
#> x23   .
#> x24   .
#> x25   .
#> x26   .
#> x27   .
#> x28   .
#> x29   .
#> x30   .
#> x31   .
#> x32   .
#> x33   .
#> x34   .
#> x35   .
#> x36   .
#> x37   .
#> x38   .
#> x39   .
#> x40   .
#> x41   .
#> x42   .
#> x43   .
#> x44   .
#> x45   .
#> x46   .
#> x47   .
#> x48   .
#> x49   .
#> x50   .
#> x51   .
#> x52   .
#> x53   .
#> x54   .
#> x55   .
#> x56   .
#> x57   .
#> x58   .
#> x59   .
#> x60   .
#> x61   .
#> x62   .
#> x63   .
#> x64   .
#> x65   .
#> x66   .
#> x67   .
#> x68   .
#> x69   .
#> x70   .
#> x71   .
#> x72   .
#> x73   .
#> x74   .
#> x75   .
#> x76   .
#> x77   .
#> x78   .
#> x79   .
#> x80   .
#> x81   .
#> x82   .
#> x83   .
#> x84   .
#> x85   .
#> x86   .
#> x87   .
#> x88   .
#> x89   .
#> x90   .
#> x91   .
#> x92   .
#> x93   .
#> x94   .
#> x95   .
#> x96   .
#> x97   .
#> x98   .
#> x99   .
#> x100  .
#> x101  .
#> x102  .
#> x103  .
#> x104  .
#> x105  .
#> x106  .
#> x107  .
#> x108  .
#> x109  .
#> x110  .
#> x111  .
#> x112  .
#> x113  .
#> x114  .
#> x115  .
#> x116  .
#> x117  .
#> x118  .
#> x119  .
#> x120  .
#> x121  .
#> x122  .
#> x123  .
#> x124  .
#> x125  .
#> x126  .
#> x127  .
#> x128  .
#> x129  .
#> x130  .
#> x131  .
#> x132  .
#> x133  .
#> x134  .
#> x135  .
#> x136  .
#> x137  .
#> x138  .
#> x139  .
#> x140  .
#> x141  .
#> x142  .
#> x143  .
#> x144  .
#> x145  .
#> x146  .
#> x147  .
#> x148  .
#> x149  .
#> x150  .
#> x151  .
#> x152  .
#> x153  .
#> x154  .
#> x155  .
#> x156  .
#> x157  .
#> x158  .
#> x159  .
#> x160  .
#> x161  .
#> x162  .
#> x163  .
#> x164  .
#> x165  .
#> x166  .
#> x167  .
#> x168  .
#> x169  .
#> x170  .
#> x171  .
#> x172  .
#> x173  .
#> x174  .
#> x175  .
#> x176  .
#> x177  .
#> x178  .
#> x179  .
#> x180  .
#> x181  .
#> x182  .
#> x183  .
#> x184  .
#> x185  .
#> x186  .
#> x187  .
#> x188  .
#> x189  .
#> x190  .
#> x191  .
#> x192  .
#> x193  .
#> x194  .
#> x195  .
#> x196  .
#> x197  .
#> x198  .
#> x199  .
#> x200  .
#> x201  .
#> x202  .
#> x203  .
#> x204  .
#> x205  .
#> x206  .
#> x207  .
#> x208  .
#> x209  .
#> x210  .
#> x211  .
#> x212  .
#> x213  .
#> x214  .
#> x215  .
#> x216  .
#> x217  .
#> x218  .
#> x219  .
#> x220  .
#> x221  .
#> x222  .
#> x223  .
#> x224  .
#> x225  .
#> x226  .
#> x227  .
#> x228  .
#> x229  .
#> x230  .
#> x231  .
#> x232  .
#> x233  .
#> x234  .
#> x235  .
#> x236  .
#> x237  .
#> x238  .
#> x239  .
#> x240  .
#> x241  .
#> x242  .
#> x243  .
#> x244  .
#> x245  .
#> x246  .
#> x247  .
#> x248  .
#> x249  .
#> x250  .
#> x251  .
#> x252  .
#> x253  .
#> x254  .
#> x255  .
#> x256  .
#> x257  .
#> x258  .
#> x259  .
#> x260  .
#> x261  .
#> x262  .
#> x263  .
#> x264  .
#> x265  .
#> x266  .
#> x267  .
#> x268  .
#> x269  .
#> x270  .
#> x271  .
#> x272  .
#> x273  .
#> x274  .
#> x275  .
#> x276  .
#> x277  .
#> x278  .
#> x279  .
#> x280  .
#> x281  .
#> x282  .
#> x283  .
#> x284  .
#> x285  .
#> x286  .
#> x287  .
#> x288  .
#> x289  .
#> x290  .
#> x291  .
#> x292  .
#> x293  .
#> x294  .
#> x295  .
#> x296  .
#> x297  .
#> x298  .
#> x299  .
#> x300  .
#> x301  .
#> x302  .
#> x303  .
#> x304  .
#> x305  .
#> x306  .
#> x307  .
#> x308  .
#> x309  .
#> x310  .
#> x311  .
#> x312  .
#> x313  .
#> x314  .
#> x315  .
#> x316  .
#> x317  .
#> x318  .
#> x319  .
#> x320  .
#> x321  .
#> x322  .
#> x323  .
#> x324  .
#> x325  .
#> x326  .
#> x327  .
#> x328  .
#> x329  .
#> x330  .
#> x331  .
#> x332  .
#> x333  .
#> x334  .
#> x335  .
#> x336  .
#> x337  .
#> x338  .
#> x339  .
#> x340  .
#> x341  .
#> x342  .
#> x343  .
#> x344  .
#> x345  .
#> x346  .
#> x347  .
#> x348  .
#> x349  .
#> x350  .
#> x351  .
#> x352  .
#> x353  .
#> x354  .
#> x355  .
#> x356  .
#> x357  .
#> x358  .
#> x359  .
#> x360  .
#> x361  .
#> x362  .
#> x363  .
#> x364  .
#> x365  .
#> x366  .
#> x367  .
#> x368  .
#> x369  .
#> x370  .
#> x371  .
#> x372  .
#> x373  .
#> x374  .
#> x375  .
#> x376  .
#> x377  .
#> x378  .
#> x379  .
#> x380  .
#> x381  .
#> x382  .
#> x383  .
#> x384  .
#> x385  .
#> x386  .
#> x387  .
#> x388  .
#> x389  .
#> x390  .
#> x391  .
#> x392  .
#> x393  .
#> x394  .
#> x395  .
#> x396  .
#> x397  .
#> x398  .
#> x399  .
#> x400  .
#> x401  .
#> x402  .
#> x403  .
#> x404  .
#> x405  .
#> x406  .
#> x407  .
#> x408  .
#> x409  .
#> x410  .
#> x411  .
#> x412  .
#> x413  .
#> x414  .
#> x415  .
#> x416  .
#> x417  .
#> x418  .
#> x419  .
#> x420  .
#> x421  .
#> x422  .
#> x423  .
#> x424  .
#> x425  .
#> x426  .
#> x427  .
#> x428  .
#> x429  .
#> x430  .
#> x431  .
#> x432  .
#> x433  .
#> x434  .
#> x435  .
#> x436  .
#> x437  .
#> x438  .
#> x439  .
#> x440  .
#> x441  .
#> x442  .
#> x443  .
#> x444  .
#> x445  .
#> x446  .
#> x447  .
#> x448  .
#> x449  .
#> x450  .
#> x451  .
#> x452  .
#> x453  .
#> x454  .
#> x455  .
#> x456  .
#> x457  .
#> x458  .
#> x459  .
#> x460  .
#> x461  .
#> x462  .
#> x463  .
#> x464  .
#> x465  .
#> x466  .
#> x467  .
#> x468  .
#> x469  .
#> x470  .
#> x471  .
#> x472  .
#> x473  .
#> x474  .
#> x475  .
#> x476  .
#> x477  .
#> x478  .
#> x479  .
#> x480  .
#> x481  .
#> x482  .
#> x483  .
#> x484  .
#> x485  .
#> x486  .
#> x487  .
#> x488  .
#> x489  .
#> x490  .
#> x491  .
#> x492  .
#> x493  .
#> x494  .
#> x495  .
#> x496  .
#> x497  .
#> x498  .
#> x499  .
#> x500  .
#> x501  .
#> x502  .
#> x503  .
#> x504  .
#> x505  .
#> x506  .
#> x507  .
#> x508  .
#> x509  .
#> x510  .
#> x511  .
#> x512  .
#> x513  .
#> x514  .
#> x515  .
#> x516  .
#> x517  .
#> x518  .
#> x519  .
#> x520  .
#> x521  .
#> x522  .
#> x523  .
#> x524  .
#> x525  .
#> x526  .
#> x527  .
#> x528  .
#> x529  .
#> x530  .
#> x531  .
#> x532  .
#> x533  .
#> x534  .
#> x535  .
#> x536  .
#> x537  .
#> x538  .
#> x539  .
#> x540  .
#> x541  .
#> x542  .
#> x543  .
#> x544  .
#> x545  .
#> x546  .
#> x547  .
#> x548  .
#> x549  .
#> x550  .
#> x551  .
#> x552  .
#> x553  .
#> x554  .
#> x555  .
#> x556  .
#> x557  .
#> x558  .
#> x559  .
#> x560  .
#> x561  .
#> x562  .
#> x563  .
#> x564  .
#> x565  .
#> x566  .
#> x567  .
#> x568  .
#> x569  .
#> x570  .
#> x571  .
#> x572  .
#> x573  .
#> x574  .
#> x575  .
#> x576  .
#> x577  .
#> x578  .
#> x579  .
#> x580  .
#> x581  .
#> x582  .
#> x583  .
#> x584  .
#> x585  .
#> x586  .
#> x587  .
#> x588  .
#> x589  .
#> x590  .
#> x591  .
#> x592  .
#> x593  .
#> x594  .
#> x595  .
#> x596  .
#> x597  .
#> x598  .
#> x599  .
#> x600  .
#> x601  .
#> x602  .
#> x603  .
#> x604  .
#> x605  .
#> x606  .
#> x607  .
#> x608  .
#> x609  .
#> x610  .
#> x611  .
#> x612  .
#> x613  .
#> x614  .
#> x615  .
#> x616  .
#> x617  .
#> x618  .
#> x619  .
#> x620  .
#> x621  .
#> x622  .
#> x623  .
#> x624  .
#> x625  .
#> x626  .
#> x627  .
#> x628  .
#> x629  .
#> x630  .
#> x631  .
#> x632  .
#> x633  .
#> x634  .
#> x635  .
#> x636  .
#> x637  .
#> x638  .
#> x639  .
#> x640  .
#> x641  .
#> x642  .
#> x643  .
#> x644  .
#> x645  .
#> x646  .
#> x647  .
#> x648  .
#> x649  .
#> x650  .
#> x651  .
#> x652  .
#> x653  .
#> x654  .
#> x655  .
#> x656  .
#> x657  .
#> x658  .
#> x659  .
#> x660  .
#> x661  .
#> x662  .
#> x663  .
#> x664  .
#> x665  .
#> x666  .
#> x667  .
#> x668  .
#> x669  .
#> x670  .
#> x671  .
#> x672  .
#> x673  .
#> x674  .
#> x675  .
#> x676  .
#> x677  .
#> x678  .
#> x679  .
#> x680  .
#> x681  .
#> x682  .
#> x683  .
#> x684  .
#> x685  .
#> x686  .
#> x687  .
#> x688  .
#> x689  .
#> x690  .
#> x691  .
#> x692  .
#> x693  .
#> x694  .
#> x695  .
#> x696  .
#> x697  .
#> x698  .
#> x699  .
#> x700  .
#> x701  .
#> x702  .
#> x703  .
#> x704  .
#> x705  .
#> x706  .
#> x707  .
#> x708  .
#> x709  .
#> x710  .
#> x711  .
#> x712  .
#> x713  .
#> x714  .
#> x715  .
#> x716  .
#> x717  .
#> x718  .
#> x719  .
#> x720  .
#> x721  .
#> x722  .
#> x723  .
#> x724  .
#> x725  .
#> x726  .
#> x727  .
#> x728  .
#> x729  .
#> x730  .
#> x731  .
#> x732  .
#> x733  .
#> x734  .
#> x735  .
#> x736  .
#> x737  .
#> x738  .
#> x739  .
#> x740  .
#> x741  .
#> x742  .
#> x743  .
#> x744  .
#> x745  .
#> x746  .
#> x747  .
#> x748  .
#> x749  .
#> x750  .
#> x751  .
#> x752  .
#> x753  .
#> x754  .
#> x755  .
#> x756  .
#> x757  .
#> x758  .
#> x759  .
#> x760  .
#> x761  .
#> x762  .
#> x763  .
#> x764  .
#> x765  .
#> x766  .
#> x767  .
#> x768  .
#> x769  .
#> x770  .
#> x771  .
#> x772  .
#> x773  .
#> x774  .
#> x775  .
#> x776  .
#> x777  .
#> x778  .
#> x779  .
#> x780  .
#> x781  .
#> x782  .
#> x783  .
#> x784  .
#> x785  .
#> x786  .
#> x787  .
#> x788  .
#> x789  .
#> x790  .
#> x791  .
#> x792  .
#> x793  .
#> x794  .
#> x795  .
#> x796  .
#> x797  .
#> x798  .
#> x799  .
#> x800  .
#> x801  .
#> x802  .
#> x803  .
#> x804  .
#> x805  .
#> x806  .
#> x807  .
#> x808  .
#> x809  .
#> x810  .
#> x811  .
#> x812  .
#> x813  .
#> x814  .
#> x815  .
#> x816  .
#> x817  .
#> x818  .
#> x819  .
#> x820  .
#> x821  .
#> x822  .
#> x823  .
#> x824  .
#> x825  .
#> x826  .
#> x827  .
#> x828  .
#> x829  .
#> x830  .
#> x831  .
#> x832  .
#> x833  .
#> x834  .
#> x835  .
#> x836  .
#> x837  .
#> x838  .
#> x839  .
#> x840  .
#> x841  .
#> x842  .
#> x843  .
#> x844  .
#> x845  .
#> x846  .
#> x847  .
#> x848  .
#> x849  .
#> x850  .
#> x851  .
#> x852  .
#> x853  .
#> x854  .
#> x855  .
#> x856  .
#> x857  .
#> x858  .
#> x859  .
#> x860  .
#> x861  .
#> x862  .
#> x863  .
#> x864  .
#> x865  .
#> x866  .
#> x867  .
#> x868  .
#> x869  .
#> x870  .
#> x871  .
#> x872  .
#> x873  .
#> x874  .
#> x875  .
#> x876  .
#> x877  .
#> x878  .
#> x879  .
#> x880  .
#> x881  .
#> x882  .
#> x883  .
#> x884  .
#> x885  .
#> x886  .
#> x887  .
#> x888  .
#> x889  .
#> x890  .
#> x891  .
#> x892  .
#> x893  .
#> x894  .
#> x895  .
#> x896  .
#> x897  .
#> x898  .
#> x899  .
#> x900  .
#> x901  .
#> x902  .
#> x903  .
#> x904  .
#> x905  .
#> x906  .
#> x907  .
#> x908  .
#> x909  .
#> x910  .
#> x911  .
#> x912  .
#> x913  .
#> x914  .
#> x915  .
#> x916  .
#> x917  .
#> x918  .
#> x919  .
#> x920  .
#> x921  .
#> x922  .
#> x923  .
#> x924  .
#> x925  .
#> x926  .
#> x927  .
#> x928  .
#> x929  .
#> x930  .
#> x931  .
#> x932  .
#> x933  .
#> x934  .
#> x935  .
#> x936  .
#> x937  .
#> x938  .
#> x939  .
#> x940  .
#> x941  .
#> x942  .
#> x943  .
#> x944  .
#> x945  .
#> x946  .
#> x947  .
#> x948  .
#> x949  .
#> x950  .
#> x951  .
#> x952  .
#> x953  .
#> x954  .
#> x955  .
#> x956  .
#> x957  .
#> x958  .
#> x959  .
#> x960  .
#> x961  .
#> x962  .
#> x963  .
#> x964  .
#> x965  .
#> x966  .
#> x967  .
#> x968  .
#> x969  .
#> x970  .
#> x971  .
#> x972  .
#> x973  .
#> x974  .
#> x975  .
#> x976  .
#> x977  .
#> x978  .
#> x979  .
#> x980  .
#> x981  .
#> x982  .
#> x983  .
#> x984  .
#> x985  .
#> x986  .
#> x987  .
#> x988  .
#> x989  .
#> x990  .
#> x991  .
#> x992  .
#> x993  .
#> x994  .
#> x995  .
#> x996  .
#> x997  .
#> x998  .
#> x999  .
#> x1000 .
#> 
#> $intercept
#> [1] -51.30934
#> 
#> $support.size
#> [1] 0
#> 
#> $support.vars
#> character(0)
#> 
#> $support.beta
#> numeric(0)
#> 
#> $dev
#> [1] 61176.56
#> 
#> $tune.value
#> [1] 1102.152
#> 

################  Formula interface  ################
data("trim32")
abess_fit <- abess(y ~ ., data = trim32)
abess_fit

#> Call:
#> abess.formula(formula = y ~ ., data = trim32)
#> 
#>    support.size         dev       GIC
#> 1             0 0.020738622 -465.0909
#> 2             1 0.008176953 -567.0402
#> 3             2 0.006186442 -590.7832
#> 4             3 0.004791180 -611.7211
#> 5             4 0.004418203 -611.7142
#> 6             5 0.003928650 -616.0745
#> 7             6 0.003616324 -616.2830
#> 8             7 0.003363741 -615.2394
#> 9             8 0.003169559 -612.6426
#> 10            9 0.002856742 -615.3798
#> 11           10 0.002594231 -617.2147
#> 12           11 0.002375152 -618.0700
#> 13           12 0.002274280 -613.5456