abess: Fast Best Subset Selection

Extremely efficient toolkit for solving the best subset selection problem <arXiv:2110.09697>. This package is its R interface. The package implements and generalizes algorithms designed in <doi:10.1073/pnas.2014241117> that exploits a novel sequencing-and-splicing technique to guarantee exact support recovery and globally optimal solution in polynomial times for linear model. It also supports best subset selection for logistic regression, Poisson regression, Cox proportional hazard model, Gamma regression, multiple-response regression, multinomial logistic regression, (sequential) principal component analysis, and robust principal component analysis. The other valuable features such as the best subset of group selection <arXiv:2104.12576> and sure independence screening <doi:10.1111/j.1467-9868.2008.00674.x> are also provided.

See also

Author

Maintainer: Jin Zhu zhuj37@mail2.sysu.edu.cn (ORCID)

Authors:

Liyuan Hu huly5@mail2.sysu.edu.cn

Junhao Huang huangjh256@mail2.sysu.edu.cn

Kangkang Jiang jiangkk3@mail2.sysu.edu.cn

Yanhang Zhang zhangyh98@ruc.edu.cn

Zezhi Wang homura@mail.ustc.edu.cn

Borui Tang tangborui@mail.ustc.edu.cn

Shiyun Lin shiyunlin@stu.pku.edu.cn

Junxian Zhu adaizjx@163.com

Canhong Wen wencanhong@gmail.com

Heping Zhang heping.zhang@yale.edu (ORCID)

Xueqin Wang wangxq20@ustc.edu.cn (ORCID)

Other contributors:

spectra contributors (Spectra implementation) [copyright holder]