Title

Title

abessbmn(
  x,
  tune.type = c("gic", "bic", "cv", "ebic", "aic"),
  weight = NULL,
  c.max = 5,
  support.size = NULL,
  gs.range = NULL,
  always.include = NULL,
  splicing.type = 0,
  max.splicing.iter = 20,
  warm.start = TRUE,
  nfolds = 5,
  foldid = NULL,
  newton.thresh = 1e-06,
  max.newton.iter = 500,
  ic.scale = 1,
  num.threads = 0,
  seed = 1,
  graph.threshold = 0,
  ...
)

Arguments

x
Input matrix, of dimension \(n \times p\); each row is an observation vector and each column is a predictor/feature/variable. Can be in sparse matrix format (inherit from class "dgCMatrix" in package Matrix).

tune.type
The type of criterion for choosing the support size. Available options are "gic", "ebic", "bic", "aic" and "cv". Default is "gic".

weight
Observation weights. When weight = NULL, we set weight = 1 for each observation as default.

c.max
an integer splicing size. Default is: c.max = 2.

support.size
An integer vector representing the alternative support sizes. Only used for tune.path = "sequence". Default is 0:min(n, round(n/(log(log(n))log(p)))).

gs.range
A integer vector with two elements. The first element is the minimum model size considered by golden-section, the later one is the maximum one. Default is gs.range = c(1, min(n, round(n/(log(log(n))log(p))))). Not available now.

always.include
An integer vector containing the indexes of variables that should always be included in the model.

splicing.type
Optional type for splicing. If splicing.type = 1, the number of variables to be spliced is c.max, ..., 1; if splicing.type = 2, the number of variables to be spliced is c.max, c.max/2, ..., 1. (Default: splicing.type = 2.)

max.splicing.iter
The maximum number of performing splicing algorithm. In most of the case, only a few times of splicing iteration can guarantee the convergence. Default is max.splicing.iter = 20.

warm.start
Whether to use the last solution as a warm start. Default is warm.start = TRUE.

nfolds
The number of folds in cross-validation. Default is nfolds = 5.

newton.thresh
a numeric value for controlling positive convergence tolerance. The Newton's iterations converge when \(|dev - dev_{old}|/(|dev| + 0.1)<\) newton.thresh.

max.newton.iter
a integer giving the maximal number of Newton's iteration iterations. Default is max.newton.iter = 10 if newton = "exact", and max.newton.iter = 60 if newton = "approx".

ic.scale
A non-negative value used for multiplying the penalty term in information criterion. Default: ic.scale = 1.

num.threads
An integer decide the number of threads to be concurrently used for cross-validation (i.e., tune.type = "cv"). If num.threads = 0, then all of available cores will be used. Default: num.threads = 0.

seed
Seed to be used to divide the sample into cross-validation folds. Default is seed = 1.

...
further arguments to be passed to or from methods.

Value

Examples

library(abess)