Adaptive best subset selection for principal component analysis

Adaptive best subset selection for principal component analysis

abesspca(
  x,
  type = c("predictor", "gram"),
  cor = FALSE,
  support.size = NULL,
  support.num = NULL,
  c.max = 2,
  lambda = 0,
  always.include = NULL,
  group.index = NULL,
  splicing.type = 1,
  max.splicing.iter = 20,
  warm.start = TRUE,
  ...
)

Arguments

x
A matrix object. It can be either the predictor matrix where each row is an observation and each column is a predictor or the sample covariance/correlation matrix.

type
If type = "predictor", x is considered as the predictor matrix. If type = "gram", x is considered as a sample covariance or correlation matrix.

cor
A logical value. If cor = TRUE, perform PCA on the correlation matrix; otherwise, the covariance matrix. This option is available only if type = "predictor". Default: cor = FALSE.

support.size
An integer vector representing the alternative support sizes.

support.num
A integer specifies the number of support size to be consider. This arguments is ignored if support.size is supplied. Default: support.num = min(ncol(x), 100) if group.index = NULL; otherwise, support.num = min(length(unique(group.index)), 100).

c.max
an integer splicing size. Default is: c.max = 2.

lambda
A single lambda value for regularized best subset selection. Default is 0.

always.include
An integer vector containing the indexes of variables that should always be included in the model.

group.index
A vector of integers indicating the which group each variable is in. For variables in the same group, they should be located in adjacent columns of x and their corresponding index in group.index should be the same. Denote the first group as 1, the second 2, etc. If you do not fit a model with a group structure, please set group.index = NULL (the default).

splicing.type
Optional type for splicing. If splicing.type = 1, the number of variables to be spliced is c.max, ..., 1; if splicing.type = 2, the number of variables to be spliced is c.max, c.max/2, ..., 1. Default: splicing.type = 1.

max.splicing.iter
The maximum number of performing splicing algorithm. In most of the case, only a few times of splicing iteration can guarantee the convergence. Default is max.splicing.iter = 20.

warm.start
Whether to use the last solution as a warm start. Default is warm.start = TRUE.

...
further arguments to be passed to or from methods.

Value

A S3 abesspca class object, which is a list with the following components:

loadings

A \(p\)-by-length(support.size) loading matrix of sparse principal components (PC), where each row is a variable and each column is a support size;

nvars

The number of variables.

support.size

The actual support.size values used. Note that it is not necessary the same as the input if the later have non-integer values or duplicated values.

ev

A vector with size length(support.size). It records the explained variance at each support size.

pev

A vector with the same length as ev. It records the percentage of explained variance at each support size.

var.all

Total variance of the explained by first principal component.

call

The original call to abess.

Details

Adaptive best subset selection for principal component analysis aim to solve the non-convex optimization problem: $$\arg\max_{v} v^\top \Sigma v, s.t.\quad v^\top v=1, \|v\|_0 \leq s, $$ where \(s\) is support size. A generic splicing technique is implemented to solve this problem. By exploiting the warm-start initialization, the non-convex optimization problem at different support size (specified by support.size) can be efficiently solved.

Author

Jin Zhu, Junxian Zhu, Ruihuang Liu, Xueqin Wang

Examples

library(abess)

## predictor matrix input:
head(USArrests)  

#>            Murder Assault UrbanPop Rape
#> Alabama      13.2     236       58 21.2
#> Alaska       10.0     263       48 44.5
#> Arizona       8.1     294       80 31.0
#> Arkansas      8.8     190       50 19.5
#> California    9.0     276       91 40.6
#> Colorado      7.9     204       78 38.7

pca_fit <- abesspca(USArrests)
str(pca_fit)

#> List of 8
#>  $ lambda      : num 0
#>  $ loadings    :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:10] 1 1 3 1 2 3 0 1 2 3
#>   .. ..@ p       : int [1:5] 0 1 3 6 10
#>   .. ..@ Dim     : int [1:2] 4 4
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : chr [1:4] "Murder" "Assault" "UrbanPop" "Rape"
#>   .. .. ..$ : chr [1:4] "1" "2" "3" "4"
#>   .. ..@ x       : num [1:10] 1 -0.9972 -0.0751 -0.9961 -0.0464 ...
#>   .. ..@ factors : list()
#>  $ ev          : num [1:4] 6945 6984 6999 7011
#>  $ nvars       : int 4
#>  $ support.size: num [1:4] 1 2 3 4
#>  $ pev         : num [1:4] 0.991 0.996 0.998 1
#>  $ var.all     : num 7011
#>  $ call        : language abesspca(x = USArrests)
#>  - attr(*, "class")= chr "abesspca"

## covariance matrix input:
pca_fit <- abesspca(stats::cov(USArrests), type = "gram")
str(pca_fit)

#> List of 8
#>  $ lambda      : num 0
#>  $ loadings    :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:10] 1 1 3 1 2 3 0 1 2 3
#>   .. ..@ p       : int [1:5] 0 1 3 6 10
#>   .. ..@ Dim     : int [1:2] 4 4
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : chr [1:4] "Murder" "Assault" "UrbanPop" "Rape"
#>   .. .. ..$ : chr [1:4] "1" "2" "3" "4"
#>   .. ..@ x       : num [1:10] 1 -0.9972 -0.0751 -0.9961 -0.0464 ...
#>   .. ..@ factors : list()
#>  $ ev          : num [1:4] 6945 6984 6999 7011
#>  $ nvars       : int 4
#>  $ support.size: num [1:4] 1 2 3 4
#>  $ pev         : num [1:4] 0.991 0.996 0.998 1
#>  $ var.all     : num 7011
#>  $ call        : language abesspca(x = stats::cov(USArrests), type = "gram")
#>  - attr(*, "class")= chr "abesspca"

## robust covariance matrix input:
rob_cov <- MASS::cov.rob(USArrests)[["cov"]]
rob_cov <- (rob_cov + t(rob_cov)) / 2
pca_fit <- abesspca(rob_cov, type = "gram")
str(pca_fit)

#> List of 8
#>  $ lambda      : num 0
#>  $ loadings    :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:10] 1 1 3 1 2 3 0 1 2 3
#>   .. ..@ p       : int [1:5] 0 1 3 6 10
#>   .. ..@ Dim     : int [1:2] 4 4
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : chr [1:4] "Murder" "Assault" "UrbanPop" "Rape"
#>   .. .. ..$ : chr [1:4] "1" "2" "3" "4"
#>   .. ..@ x       : num [1:10] 1 -0.997 -0.0772 -0.9948 -0.0658 ...
#>   .. ..@ factors : list()
#>  $ ev          : num [1:4] 6575 6614 6642 6655
#>  $ nvars       : int 4
#>  $ support.size: num [1:4] 1 2 3 4
#>  $ pev         : num [1:4] 0.988 0.994 0.998 1
#>  $ var.all     : num 6655
#>  $ call        : language abesspca(x = rob_cov, type = "gram")
#>  - attr(*, "class")= chr "abesspca"