Extract one model from a fitted "abess" object.

Extract the fixed-support-size model's information such as the selected predictors, coefficient estimation, and so on.

extract(object, support.size = NULL, ...)

# S3 method for abess
extract(object, support.size = NULL, ...)

Arguments

object

An "abess" project.

support.size

An integer value specifies the model size fitted at given support.size. If support.size = NULL, then the model with the best tuning value would be returned. Default: support.size = NULL.

...

Other arguments.

Value

A list object including the following components:

beta

A \(p\)-by-1 matrix of sparse matrix, stored in column format.

intercept

The fitted intercept value.

support.size

The support.size used in the function.

support.beta

The support.size-length vector of fitted coefficients on the support set.

support.vars

The character vector gives variables in the support set.

tune.value

The tuning value of the model.

dev

The deviance of the model.

See also