Generate a dataset with independent observations from binary markov network

Generate a dataset with independent observations from binary markov network

generate.bmn.data(
  n,
  p,
  type = 1,
  seed = NULL,
  graph.seed = NULL,
  beta = 0.7,
  degree = 3,
  alpha = 0.4
)

Arguments

n

p

type

seed

Value