Nodewise logistic regression for inverse Ising problem

Nodewise logistic regression for inverse Ising problem

nodewise_L0(
  x,
  max.support.size = NULL,
  tune.type = "cv",
  foldid = NULL,
  support.size = NULL,
  graph.threshold = 0
)

Arguments

x
Input matrix, of dimension \(n \times p\); each row is an observation vector and each column is a predictor/feature/variable. Can be in sparse matrix format (inherit from class "dgCMatrix" in package Matrix).

max.support.size

tune.type
The type of criterion for choosing the support size. Available options are "gic", "ebic", "bic", "aic" and "cv". Default is "gic".

support.size
An integer vector representing the alternative support sizes. Only used for tune.path = "sequence". Default is 0:min(n, round(n/(log(log(n))log(p)))).

Value

Examples

p <- 16
n <- 1e5 / 2
library(abess)
train <- generate.bmn.freq.data(n, p, type = 10, graph.seed = 1, seed = 1)
res <- nodewise_L0(train[["data"]], tune.type = "gic", 
                   max.support.size = 4, support.size = rep(4, p))
all((res[[1]] != 0) == (train[["theta"]] != 0))
#> [1] TRUE

valid <- generate.bmn.freq.data(n, p, type = 10, graph.seed = 1, seed = 2)
all(train[["theta"]] == valid[["theta"]])
#> [1] FALSE
x <- rbind(train[["data"]], valid[["data"]])
res <- nodewise_L0(x, tune.type = "cv")
all((res[[1]] != 0) == (train[["theta"]] != 0))
#> [1] FALSE